+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/EnrollmentEditEventPage.component.js
@@ -56,6 +56,12 @@ export const EnrollmentEditEventPageComponent = ({
     onSaveAssigneeError,
     onDeleteTrackedEntitySuccess,
     onAccessLostFromTransfer,
+    onNavigateToEvent,
+    onDeleteEvent,
+    onDeleteEventRelationship,
+    onUpdateOrAddEnrollmentEvents,
+    onUpdateEnrollmentEventsSuccess,
+    onUpdateEnrollmentEventsError,
 }: PlainProps) => (
     <OrgUnitFetcher orgUnitId={orgUnitId}>
         <TopBar
@@ -119,6 +125,12 @@ export const EnrollmentEditEventPageComponent = ({
             onAccessLostFromTransfer={onAccessLostFromTransfer}
             feedbackEmptyText={i18n.t('No feedback for this event yet')}
             indicatorEmptyText={i18n.t('No indicator output for this event yet')}
+            onNavigateToEvent={onNavigateToEvent}
+            onDeleteEvent={onDeleteEvent}
+            onDeleteEventRelationship={onDeleteEventRelationship}
+            onUpdateOrAddEnrollmentEvents={onUpdateOrAddEnrollmentEvents}
+            onUpdateEnrollmentEventsSuccess={onUpdateEnrollmentEventsSuccess}
+            onUpdateEnrollmentEventsError={onUpdateEnrollmentEventsError}
         />
         <NoticeBox formId={`${dataEntryIds.ENROLLMENT_EVENT}-${mode}`} />
     </OrgUnitFetcher>