+++ src/core_modules/capture-core/components/WidgetTwoEventWorkspace/OverflowMenu/Modal/UnlinkAndDeleteModal.types.js
@@ -3,4 +3,7 @@ export type Props = {|
         setOpenModal: (open: boolean) => void,
         eventId: string,
         originEventId: string,
+        relationshipId: string,
+        onDeleteEvent?: (eventId: string) => void,
+        onDeleteEventRelationship?: (relationshipId: string) => void,
 |};